+++ test/unit/constants.js
@@ -1,7 +1,7 @@
 /**
     Set of constants focused on MRS.Core automated tests
 **/
-angular.module("MRS.i18n").constant("$mrsi18nConfig", {
+angular.module("MRS.App.i18n").constant("$mrsappi18nConfig", {
 	"resources": {
 		"path": "../test/unit/resources/"
 	}