+++ CHANGELOG.md
@@ -1,6 +1,18 @@
-TlanJS - i18n v0.4.x release notes
+MRS App i18n release notes
 ==============================================
 
-## v0.4.2 - Changelog
+## latest
+
+This release changes the module namespace from MRS to MRS.App.
+
+### Changes
+* Added interceptor languageInterceptor
+* Added banner to minified files
+
+### Break changes
+* Namespace from MRS.i18n to MRS.App.i18n
+* Service from translate to i18nTranslate
+
+## v0.4.2
 
 This release represents the end of the TFS as official repository for the old MRS.i18n and the birth of TlanJS_i18n on Github.